backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.6.0
aiohttp>=3.9.3
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
requests>=2.32.0